Fix: params error 수정
