+++ src/app/api/stores/route.ts
@@ -3,7 +3,7 @@ import prisma from "@/db";
 import axios from "axios";
 
 import { getServerSession } from "next-auth";
-import { authOptions } from "@/app/api/auth/[...nextauth]/route";
+import { authOptions } from "@/app/utils/authOptions";
 
 export async function GET(req: Request) {
   const { searchParams } = new URL(req.url);
